Use new upgrade test code.
