+++ test/go.mod
@@ -3,7 +3,7 @@ module github.com/gruntwork.io/package-terraform-utilities/test
 go 1.14
 
 require (
-	github.com/gruntwork-io/module-ci/test/upgrades v0.0.0-20220706191102-c35449193007
+	github.com/gruntwork-io/module-ci/test/upgrades v0.0.0-20220808203123-0281ca0983d1
 	github.com/gruntwork-io/terratest v0.40.14
 	github.com/jinzhu/copier v0.1.0 // indirect
 	github.com/stretchr/testify v1.7.1
